fe performance and polishing
