+++ lib/hooks/supabase/useInvoice.ts
@@ -62,6 +62,7 @@ export const useInvoice = () => {
           headers: {
             'Content-Type': 'application/json',
           },
+          cache: 'no-cache',
         },
       );
